OPHVKTKEH-55: ilmoittautumisen alustus

+++ backend/vkt/src/main/java/fi/oph/vkt/api/dto/PublicExamEventDTO.java
@@ -12,6 +12,6 @@ public record PublicExamEventDTO(
   @NonNull LocalDate date,
   @NonNull LocalDate registrationCloses,
   @NonNull Long participants,
-  @NonNull Integer maxParticipants,
+  @NonNull Long maxParticipants,
   @NonNull Boolean hasCongestion
 ) {}
