+++ backend/vkt/src/main/java/fi/oph/vkt/repository/PublicExamEventProjection.java
@@ -9,5 +9,5 @@ public record PublicExamEventProjection(
   LocalDate date,
   LocalDate registrationCloses,
   long participants,
-  int maxParticipants
+  long maxParticipants
 ) {}
